requirements.txt
Flask
Flask-Script
Flask-Session
flask-restful
Flask-PyMongo
flask_restful_swagger
vk
facebook-sdk
